Fixes #73
In UK (possibly "outside of everywhere but USA") the attempts to get book's ASIN via '/buttons/glide/{id}' links fails - the ASIN is not in there.  We can fix that by looking in the tooltip file we already have - it is often there, and this saves a fair number of network operations.  We keep the old method as a backup, as in the USA at least this might help.

Also fixes a few other minor things:
- removes unnecessary duplication in __init__
- adds missing initialisation in __init__ (was getting an assert tryign to check if author_recommendations was not None, which only works if 'author_recommendations' exists)
- remove 'current_book_asin' - was doing a network op, and then doing nothing with the result